requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
filelock==3.16.1
fsspec==2024.10.0
idna==3.10
Jinja2==3.1.4
llvmlite==0.43.0
logging==0.4.9.6
MarkupSafe==3.0.2
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
openai-whisper==20240930
pathlib==1.0.1
pillow==11.0.0
regex==2024.11.6
requests==2.32.3
sympy==1.13.1
tiktoken==0.8.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
